Build-constrain the Prometheus memory map metrics
Windows does not have the concept of a count of mapped memory areas.

Non-linux operating systems do not use the /sys pseudo-filesystem.

Constrain `mustRegisterMemoryMapMetrics` to linux-only builds.
Other platforms use a no-op function.
